requirements.txt
babel==2.16.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.6
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.9.0
docx==0.2.4
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.18.0
et-xmlfile==1.1.0
evdev==1.7.1
fonttools==4.53.1
fpdf2==2.7.9
fuzzywuzzy==0.18.0
html5lib==1.1
idna==3.8
inquirerpy==0.3.4
Jinja2==3.1.4
Levenshtein==0.25.1
lml==0.1.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
mss==9.0.2
numpy==2.1.0
odfpy==1.4.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pfzy==0.3.4
pillow==10.4.0
prompt_toolkit==3.0.47
PyAutoGUI==0.9.54
pycparser==2.22
pydyf==0.11.0
pyexcel==0.7.0
pyexcel-io==0.6.6
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pynput==1.7.7
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
pyphen==0.16.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-Levenshtein==0.25.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
rapidfuzz==3.9.7
reportlab==4.2.2
requests==2.32.3
rich==13.8.0
schedule==1.2.2
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
tabula-py==2.9.3
texttable==1.7.0
tinycss2==1.3.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
watchdog==5.0.2
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
xlsx2csv==0.8.3
XlsxWriter==3.2.0
zopfli==0.2.3
